Use `source-inplace` instead of `source`
hdevtools uses the directory that the source file is found in to locate cabal sandboxes and (in version 1.2.1) to locate `stack.yaml` files.

Using `source` makes hdevtools unable to find those, since it then only has a temporary directory as the location of the file.

On the other hand, using a filename that doesn't correspond exactly to the name of the module is fine with hdevtools, and doesn't generate any spurious warnings. Therefore, `source-inplace` is the better choice.